app/requirements.txt
click==8.1.3
clickhouse-driver==0.2.6
dash==2.9.3
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.3.2
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.3
packaging==23.1
pandas==2.0.1
plotly==5.14.1
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
SQLAlchemy==2.0.13
tenacity==8.2.2
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
Werkzeug==2.3.4
